Unite le modifiche per la scrittura dei file su disco. Risolto un problema di scrittura.

+++ src/model/JSONFiles/JSONFileWriter.java
@@ -72,8 +72,8 @@ public class JSONFileWriter {
 			
 			JSONObject nodeObj = new JSONObject() ;
 			
-			nodeObj.put("x", node.getElement().getxPos()) ;
-			nodeObj.put("y", node.getElement().getyPos()) ;
+			nodeObj.put("x", (int) node.getElement().getxPos()) ;
+			nodeObj.put("y", (int) node.getElement().getyPos()) ;
 			
 			verticesListObj.put(Integer.toString(node.getElement().getIndex()), nodeObj) ;
